fix: move `getWithAppKeyAndToken` function to its own module
